requirements.txt
aiogram==2.2
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
Babel==2.14.0
certifi==2023.11.17
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.6
multidict==6.0.4
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.2.0
yarl==1.9.4
